Rename DVD_dataset file to DVDDataset for consistency in naming
