+++ codes/data/__init__.py
@@ -3,8 +3,8 @@ import logging
 
 import torch.utils.data
 
-from codes.data.DVD_dataset import DVDDataset
-from codes.data.DVD_dataset import DVDIDataset
+from codes.data.DVDDataset import DVDDataset
+from codes.data.DVDDataset import DVDIDataset
 from codes.data.LRHRC_dataset import LRHRDataset as LRHRCDataset
 from codes.data.LRHROTF_dataset import LRHRDataset as LRHROTFDataset
 from codes.data.LRHRPBR_dataset import LRHRDataset as LRHRPBRDataset
